add more debug

+++ src/kafka/records.rs
@@ -59,8 +59,8 @@ impl RecordsBatch {
 pub struct KafkaRecord {
     pub length: i8,
     pub attributes: i8,
-    pub timestamp_delta: i8,
-    pub offset_delta: i8,
+    pub timestamp_delta: u64,
+    pub offset_delta: i32,
     pub key: Vec<u8>,
     pub value: Vec<u8>,
     pub headers: Vec<KafkaRecordHeader>,
